requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0

google-api-python-client>=2.139.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
